Remove balance-form component and update account features
The balance-form.component has been removed as it is no longer used in the application. Minor adjustments are made in the login-form component where unnecessary console logs are removed for better code quality. A new method for adding an account is added in account.service.ts for future growth and scalability. The account.component has been updated to import the add-account component instead of the removed balance-form component. A toggle for a dialog of adding account is added in account.component.ts as a new feature. The account component's UI has been updated for displaying sum of all account's balance instead of individual's account. Finally, the account path routing has been updated for adding the route to the add-account component.