Remove 'title' attribute from ICalendarDay model
The 'title' attribute, which was previously used to store date value as string, has been removed from ICalendarDay model. This field was unnecessary as 'date' attribute already stores the date. This also affected the calendar-state.service where 'title' initialization was removed. This change is done to improve data redundancy and performance.