Update calendar UI for better usability and appearance
The calendar feature has undergone several changes to improve usability and enhance visual appeal. The `day-card` component was updated to a card-like design and the `calendar-table` was changed to a grid layout. Also, minor changes were added to the hover effect on `day-card-item` component. The changes also included adding day names constants, and updating component styles for better responsiveness and visual consistency.

The improvement was primarily done for the flexibility of card layout and allow better control over CSS styling. The changes are expected to improve the look and feel of the application, making it a more optimal experience for users.