Minor UI Detail change - day-card
Tweaked the styling on day-card and calendar-table components. Specifically added box-shadow on hover state in day-card to increase depth perception which provides an easier visual tracking of user's interaction. Also modified padding and background color for calendar-table component to enhance readability and adjusted the width of the month-button for aesthetic conformity. The changes were made to improve UX by making the interface more visually appealing."