"Refactor day-card and day-card-item component structure and styles
HTML structure and SCSS styles for day-card and day-card-item components have been refactored for enhanced readability and maintainability of code. Changes include renaming of div classes for better semantics and adjustments to styles. The addition of logos to the items also improves user experience by giving a graphical cue for income and expenses."