"Update Add Budget form and functionality"
Updated the "Add Budget" form and improved its usability. Changed the function triggered on click from "show()" to "showAddBudgetDialog()" on the calendar day component. The form was revamped with dropdowns for multiple options like transaction type, frequency, and account. Also, added a color picker and note textarea. Moreover, the form now supports the creation of recurring budgets. These changes were necessary to enhance the user experience and ease of use. Additionally, created a model 'iBudgetToCreate.ts' for the budget to be created.