"Replace dropdown with autocomplete for category selection"
This change replaces the dropdown selection for budget categories within the add budget form with an autocomplete field. The new field connects to a CategoryService to fetch the category suggestions dynamically, and the typed-in category names that do not match existing categories are also accepted.

The HTTP response from the server-side CategoryController has been refactored as well to deliver the category data directly. This change provides a more flexible and user-friendly way to select a category, and reduces the amount of transmitted data by eliminating unnecessary extra information in the HTTP response.