Add budget range retrieval and transaction totals calculation
A new function was added to the budget service to get budgets by date range. This function uses a new API endpoint, the path of which has been added to api-paths.constants. The calendar-day component now calculates and stores totals for income and expense transactions and logs these values to the console. This information will give users a better understanding of their income and expenses for each day.

Also, the "Microsoft.AspNetCore.OpenApi" package in the Server project was updated to version 7.0.9 for better API documentation and improvements made in generate-days service for better performance.