Update budget and authService services
The changes include refactoring the budget and authService services and updating the username function in the AuthController.cs file. The refactoring was done to render the AuthService reusable across other parts of the application and to normalize the username in the AuthenticationResponse to better adhere to naming standards. Debug logs and console.log statements are removed to clean up the code, getCategories and getBudgetsByRange services are refracted to maintain business logic in the service layer. Sidebar and header navigation components were updated to integrate an improved menu and user interaction. Display style changes were introduced in CSS to better fit the elements within the viewport height.

Deleted mobile-budget-page, budget-form and transaction-type-color.pipe as these were causing redundancies in the code.

Fixed the header and thread components to only appear when user is authenticated by the server.

Creates userName signal in login.service to store the username for later use in the app. When the login() function is called, it subscribes to the httpClient.post() Observable, which in turn executes the HTTP POST request to the server with the body of type IUserForAuthenticationDto. If the server response (res) is successful and includes a token,  the token is stored in the localStorage, and if a username is included in the server response, it is logged in the userName signal. If the response from the server is not successful or does not include a token, an error is logged.